+++ DBModel/Models/Packaging.cs
@@ -10,6 +10,6 @@ namespace DBModel.Models
         public int PackagingId { get; set; }
         [MaxLength(200)]
         public string PackagingType { get; set; }
-        public int Tara { get; set; }
+        public decimal Tara { get; set; }
     }
 }
